fix: removed special handling of yoga react component
Now the component will be require with the requireComponent from worflow-react.
