+++ packages/workflow-layout-yoga/flows/CenterExample.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import render, { Workspace, requireComponent } from 'workflow-react';
-import { Flex } from '../components';
 
+const { Flex } = requireComponent('workflow-layout-yoga');
 const { Browser } = requireComponent('workflow-apps-defaults');
 
 export default render(
